Changes to support testing entirely within the UDx container environment
Add vsdk-vertica --- a script that starts a Vertica in a container
for you to test with.

Changed Dockerfile_RPM to build a container with a runnable Vertica inside
it (this is derived from the one-node-ce Dockerfile).

entrypoint.sh can handle the situation where it is invoked to run a Vertica.