Allow the RPM to be in a different directory
Docker requires the install files to be in the current working
directory.  This change will create a hardlink to the RPM file , but if
the RPM is on a different device, it falls back to copying the rpm to
the current working directory.  Then, after docker is done with it, the
newly created link/copy is deleted.  Interrupting the process should
still remove the file if it can.