Add the --rm flag to docker run
...to cause the temporary container it creates to be removed right after
it's done.  Generally if you want a container to last, you give it a
name rather than let docker chose a random name for it.  vsdk-vertica
sets a container name, so I didn't add the --rm flag there, although I
think it might be useful as well otherwise running it a second time
might give an error.